\name{getEvidenceCodes}
\alias{getEvidenceCodes}
\title{Returns GO evidence codes.}
\description{Returns GO evidence codes.}
\references{\code{?org.Bt.egGO}}
\value{Matrix of two columns, first column with codes, second column with description of codes.}
\seealso{\code{\link{pickGO}}}
\author{Stefan McKinnon Edwards \email{stefan.hoj-edwards@agrsci.dk}}
\examples{getEvidenceCodes()}
