\name{.pickRef}
\alias{.pickRef}
\title{Secret function that does the magic for pickRefSeq.}
\usage{.pickRef(l, priorities, reduce=c("all", "first", "last"))}
\description{Secret function that does the magic for pickRefSeq.}
\details{Do not use it, use \code{\link{pickRefSeq}}!}
\value{List.}
\note{Hey, you found a secret function! Keep it that way!}
\author{Stefan McKinnon Edwards \email{stefanm.edwards@agrsci.dk}}
\seealso{\code{\link{pickRefSeq}}}
\arguments{\item{l}{List.}
\item{priorities}{How to prioritize.}
\item{reduce}{How to reduce.}}
